% Auto-generated: do not edit by hand
\name{''ReactFlickityCarousel}

\alias{''ReactFlickityCarousel}

\title{ReactFlickityCarousel component}

\description{

}

\usage{
''ReactFlickityCarousel(children=NULL, id=NULL, className=NULL, options=NULL,
value=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers. }

\item{id}{Character. }

\item{className}{Character. }

\item{options}{Named list. }

\item{value}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

